backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Letta for stateful agent
letta>=0.4.0

# HTTP client for API calls
httpx>=0.26.0
aiohttp>=3.9.0

# Google Calendar integration
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.116.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Date/time handling
python-dateutil>=2.8.2

# Environment variables
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0  # Also needed for FastAPI TestClient

# Firebase (optional - for production push notifications)
# firebase-admin>=6.3.0
